requirements.txt
beautifulsoup4==4.9.3
cbsodata==1.3.2
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dateparser==1.0.0
distro==1.5.0
idna==2.10
isoweek==1.3.3
kiwisolver==1.2.0
lxml==4.6.2
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
patsy==0.5.1
pdfminer==20191125
Pillow==8.0.1
pycryptodome==3.9.9
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.11.13
requests==2.24.0
scipy==1.5.3
six==1.15.0
soupsieve==2.0.1
statsmodels==0.12.0
tabula-py==2.2.0
tabulate==0.8.7
tzlocal==2.1
urllib3==1.25.11
